Corrected return type of "stage_no_of"

+++ Model/Plant/Utility/Science.cpp
@@ -73,7 +73,7 @@ void accumulate (float value,             //  (INPUT) value to add to array
    }
 
 //===========================================================================
-float stage_no_of (float stage_code,           //(INPUT) stage code to look up
+int stage_no_of (float stage_code,           //(INPUT) stage code to look up
                    float *stage_code_list,     //(INPUT) list of stage codes
                    int   list_size)            //(INPUT) size_of of stage code list
 //===========================================================================
